+++ common/src/main/java/com/tendai/common/source/model/Playlist.kt
@@ -2,7 +2,7 @@ package com.tendai.common.source.model
 
 
 data class Playlist(
-    var playlistId: Int = -1,
+    var playlistId: Long = -1,
     var playlistName: String = "",
     var numberOfTracks: Int = -1
 )
